Add JWT logic and serialization

+++ src/events/events.controller.ts
@@ -1,6 +1,5 @@
 import { CreateEventDto } from "./input/create-event.dto";
 import { UpdateEventDto } from "./input/update-event.dto";
-import { Event } from "./event.entity";
 import {
     Controller,
     Get,
@@ -16,27 +15,30 @@ import {
     Query,
     UsePipes,
     ValidationPipe,
+    UseGuards,
+    ForbiddenException,
+    SerializeOptions,
+    UseInterceptors,
+    ClassSerializerInterceptor,
 } from "@nestjs/common";
-import { Repository } from "typeorm";
-import { InjectRepository } from "@nestjs/typeorm";
-import { Attendee } from "./attendee.entity";
 import { EventsService } from "./event.service";
 import { ListEvents } from "./input/list.events";
+import { CurrentUser } from "src/auth/current-user.decorator";
+import { User } from "src/auth/user.entity";
+import { AuthGuardJwt } from "src/auth/auth-guard.jwt";
 
 @Controller("/events")
+@SerializeOptions({ strategy: "excludeAll" })
 export class EventsController {
     private readonly logger = new Logger(EventsController.name);
 
     constructor(
-        @InjectRepository(Event)
-        private readonly repository: Repository<Event>,
-        @InjectRepository(Attendee)
-        private readonly attendeeRepository: Repository<Attendee>,
         private readonly eventsService: EventsService
     ) {}
 
     @Get()
     @UsePipes(new ValidationPipe( { transform: true } ))
+    @UseInterceptors(ClassSerializerInterceptor)
     async findAll(@Query() filter: ListEvents) {
         this.logger.debug(`Hit the findAll route`);
         const events = await this.eventsService.getEventsWithAttendeeCountFilteredPaginated(filter, {
@@ -48,42 +50,43 @@ export class EventsController {
         return events;
     }
 
-    @Get("/practice")
-    async practice() {
-        return await this.repository.find({
-            where: { id: 3 },
-        });
-    }
-
-    @Get("/practice2")
-    async practice2() {
-        //     return await this.repository.findOne({
-        //         where: { id: 1 },
-        //         relations: ['attendees'] });
-        // }
-        //     const event = await this.repository.findOne({where: { id: 1 }});
-        //     const attendee = new Attendee();
-        //     attendee.name = "Jerry";
-        //     attendee.event = event;
-        //     await this.attendeeRepository.save(attendee);
-        //     return event;
-        // }
-    //     const event = await this.repository.findOne({ where: { id: 1 } });
-    //     const attendee = new Attendee();
-    //     attendee.name = "Jerry";
-    //     event.attendees.push(attendee);
-    //     await this.repository.save(event);
-    //     return event;
+    // @Get("/practice")
+    // async practice() {
+    //     return await this.repository.find({
+    //         where: { id: 3 },
+    //     });
     // }
 
-        return await this.repository.createQueryBuilder("e")
-        .select(["e.id", "e.name"])
-        .orderBy("e.id", "ASC")
-        .take(3)
-        .getMany();
-    }
+    // @Get("/practice2")
+    // async practice2() {
+    //     //     return await this.repository.findOne({
+    //     //         where: { id: 1 },
+    //     //         relations: ['attendees'] });
+    //     // }
+    //     //     const event = await this.repository.findOne({where: { id: 1 }});
+    //     //     const attendee = new Attendee();
+    //     //     attendee.name = "Jerry";
+    //     //     attendee.event = event;
+    //     //     await this.attendeeRepository.save(attendee);
+    //     //     return event;
+    //     // }
+    // //     const event = await this.repository.findOne({ where: { id: 1 } });
+    // //     const attendee = new Attendee();
+    // //     attendee.name = "Jerry";
+    // //     event.attendees.push(attendee);
+    // //     await this.repository.save(event);
+    // //     return event;
+    // // }
+
+    //     return await this.repository.createQueryBuilder("e")
+    //     .select(["e.id", "e.name"])
+    //     .orderBy("e.id", "ASC")
+    //     .take(3)
+    //     .getMany();
+    // }
 
     @Get(":id")
+    @UseInterceptors(ClassSerializerInterceptor)
     async findOne(@Param("id", ParseIntPipe) id) {
         const event = await this.eventsService.getEvent(id);
         if (!event) {
@@ -93,32 +96,39 @@ export class EventsController {
     }
 
     @Post()
-    async create(@Body() input: CreateEventDto) {
-        return await this.repository.save({
-            ...input,
-            when: new Date(input.when),
-        });
+    @UseGuards(AuthGuardJwt)
+    @UseInterceptors(ClassSerializerInterceptor)
+    async create(@Body() input: CreateEventDto, @CurrentUser() user: User) {
+        return await this.eventsService.createEvent(input, user);
     }
 
     @Patch(":id")
-    async update(@Param("id", ParseIntPipe) id, @Body() input: UpdateEventDto) {
-        const event = await this.repository.findOne(id);
+    @UseGuards(AuthGuardJwt)
+    @UseInterceptors(ClassSerializerInterceptor)
+    async update(@Param("id", ParseIntPipe) id, @Body() input: UpdateEventDto, @CurrentUser() user: User) {
+        const event = await this.eventsService.getEvent(id);
         if (!event) {
             throw new NotFoundException();
         }
-        return await this.repository.save({
-            ...event,
-            ...input,
-            when: input.when ? new Date(input.when) : event.when,
-        });
+
+        if (event.organizerId !== user.id) {
+            throw new ForbiddenException(null, `You are not authorized to change this event`);
+        }
+
+        return await this.eventsService.updateEvent(event, input);
     }
 
     @Delete(":id")
+    @UseGuards(AuthGuardJwt)
     @HttpCode(204)
-    async remove(@Param("id", ParseIntPipe) id) {
-        const result = await this.eventsService.deleteEvent(id);
-        if (result?.affected !== 1) {
+    async remove(@Param("id", ParseIntPipe) id, @CurrentUser() user: User) {
+        const event = await this.eventsService.getEvent(id);
+        if (!event) {
             throw new NotFoundException();
         }
+        if (event.organizerId !== user.id) {
+            throw new ForbiddenException(null, `You are not authorized to remove this event`);
+        }
+        await this.eventsService.deleteEvent(id);
     }
 }
